Move ingress splitter to internal package
