+++ go/vt/vtgate/vtgateservice/ingress.go
@@ -48,57 +48,3 @@ func IngressBytesForQuery(ctx context.Context, index int) (uint64, bool) {
 	}
 	return ingressBytes[index], true
 }
-
-// SplitIngressBytes distributes request-level ingress bytes across per-query
-// log stats using caller-provided weights. Zero-weight entries receive no bytes
-// unless all entries have zero weight, in which case the bytes are split evenly.
-//
-// The result always sums to total; any rounding remainder is assigned to the
-// last positive-weight entry.
-func SplitIngressBytes(total uint64, weights []int) []uint64 {
-	allocations := make([]uint64, len(weights))
-	if len(weights) == 0 || total == 0 {
-		return allocations
-	}
-
-	var totalWeight uint64
-	for _, weight := range weights {
-		if weight > 0 {
-			totalWeight += uint64(weight)
-		}
-	}
-	if totalWeight == 0 {
-		splitIngressBytesEvenly(total, allocations)
-		return allocations
-	}
-
-	lastPositiveWeight := -1
-	for i, weight := range weights {
-		if weight > 0 {
-			lastPositiveWeight = i
-		}
-	}
-	var allocated uint64
-	for i, weight := range weights {
-		if i == lastPositiveWeight {
-			allocations[i] = total - allocated
-			break
-		}
-		if weight > 0 {
-			allocations[i] = total * uint64(weight) / totalWeight
-		}
-		allocated += allocations[i]
-	}
-	return allocations
-}
-
-func splitIngressBytesEvenly(total uint64, allocations []uint64) {
-	base := total / uint64(len(allocations))
-	remainder := total % uint64(len(allocations))
-	for i := range allocations {
-		allocations[i] = base
-		if uint64(i) < remainder {
-			allocations[i]++
-		}
-	}
-}
